Rewritten Spline.build(), adding initializer arg
Spline.build() completely rewritten, to clean the code and adding an
"initializer" argument to specify how to generate the initial
train_points, either "random"ly or in "grid" pattern. In this last option,
if n_control_points = N**2 then the grid will be formed with NxN
points (more specifically, N is taken as round(sqrt(n_control_points), so the true number of control points could be different than
n_control_points).
Masked pixels are evaluated as if their value was the background value
(estimated as the median of im).
Also rewritten the handling of different mask specification methods,
in particular, introducing a (threshold_min, threshold_max) option.
Giving a min threshold is interesting when there are missing values
represented as 0.0 (or very low value below the pedestal) typical in
registered images, but also when handling dark landscapes in wide
shots if the landscape histogram is separated from the sky histogram.
The argument regularization_weight of tfa._solve_interpolation() is
added as a parameter of the Spline() constructor.