New Spline._apply_mask(). Cosmetic changes in bgoptimizer
New static method _apply_mask() in Spline that encapsulates mask
generation when given as a threshold tuple. Useful also for previewing
the mask.

Cosmetic changes on bgoptimizer for better output. Also changed
plt.imshow() invocations to ensure (h,w) shape instead of (h,w,1),
that is not allowed in matplotlib (it shows error on recent versions).