Code refactoring.
bgoptimizer is now a command line script.
Preprocessing and postprocessing are encapsulated on the ImageProcessor
class (imageprocessor.py)
Background model training and interpolation to full resolution is now
encapsulated on BgModel class (bgmodel.py)
Tested on RGB images also, although the loss function may not be the
optimal for this case (it may be better just to run the optimization for each
channel separately).